WIP: Make Avro DatumWriter implementation configurable
Currently, the logic for deciding which avro `DatumWriter` to use is not
open to extension.

This pull requests adds a new configuration option,
"generic.avro.writer.class", for specifying which `DatumWriter`
implementation to use when writing _generic_ records.

This configuration option is respected by `KafkaAvroSerializer` which
will use instances of the specified `DatumWriter` implementation when
serializing. Default is `GenericDatumWriter`.