Switch serializer config classes to use AbstractConfig from confluent-common instead of from Kafka.
As part of b42bde7d, we switched to using AbstractConfig instead of duplicating
a bunch of config parsing across different serializer/encoder
interfaces. However, it was using Kafka's AbstractConfig, which is not part of
the public interface where compatibility is maintained. This could cause issues
if the serializer is used with a newer version of Kafka where those classes have
changed (as they did recently after 0.8.2.1).

The drawback to this is that we have an extra dependency for
serializers. However, we can ensure compatibility with confluent-common and it
is a pretty tiny dependency.