Split packaging of schema registry and serde tools
This commit splits the packaging of schema registry (server and client)
from the packaging of Avro serializer, Avro converter, and JSON
serializer.  This allows us to have two different /usr/share/java/*
directories (`schema-registry` and `kafka-serde-tools`), with each
directory containing the proper set of dependencies -- which didn't work
out of the box before and was the reason we eventually packaged
everything into a single directory in commit 5d5d62e.