CC-44: Include version when deserializing Kafka Connect data.
We were already including the version from the schema registry in the Avro
schema, but in a field that could potentially conflict with other systems and it
was not propagated to the Connect data. This ensures we use the Connect schema
version included by the source if it exists, then falls back on the schema
registry version.