Encode null values for schemaless array entries and map keys and values as an Anything record.
This is requred because the catch-all Anything schema cannot be a union since it
cannot be named, so null values cannot be encoded directly. If an entire value
is null we can use null directly, but for internal values (array entries, map
keys or values) we need to encode a null as an Anything struct with no non-null
fields.