Handle Optional Values within Records
Summary:
This change allows for the use of optional fields, especially optional
strings within a record. Prior to the change the only null check on
value is in the public toConnectData method. This isn't sufficient when
handling Avro Records. It is common to have a Record that has optional
types within it.