CLIENTS-488 Updated distributed worker config for internal topic creation
Updated the distributed worker configuration used in the examples to include the new properties used when Connect creates the internal storage topics. The complicating factor is that the examples should work when run with a single broker, but the default replication factor for the topics is 3, which requires at least three brokers. Therefore, like the AK example configuration, the replication factor is set to 1 for the three topics and comments are used to describe why that is done and what should be used for production.