New (key|value).multi.type option for Avro serialization
In some situations, an application needs to store events of several
different types in the same Kafka topic. In particular, when developing
a data model in an Event Sourcing style, you might have several kinds of
event that affect the state of an entity. For example, for a customer
entity there may be customerCreated, customerAddressChanged,
customerEnquiryReceived, customerInvoicePaid, etc. events, and the
application may require that those events are always read in the same
order. Thus, they need to go in the same Kafka partition (to maintain
ordering).

The Avro schema registry currently assumes a 1:1 mapping between Kafka
topics and Avro schemas, making it difficult to support scenarios like
the one above. Users who want several event types in the same topic
currently either have to put them in one big Avro union (which works,
but gets unwieldy very quickly), or turn off the registry's schema
compatibility checking (which would be unfortunate, since the
compatibility check is very valuable).

This patch introduces two new boolean config settings, key.multi.type
and value.multi.type. When set to true, they allow the key (or value,
respectively) of a message to be *any* Avro record type. The schema of
the type is stored in the schema registry as usual; however, instead of
using "<topic>-key" or "<topic>-value" as subject, the fully-qualified
name of the record type is used as subject.

This has the effect that a Kafka producer will happily accept any
mixture of Avro record types and publish them to the same topic. Since
the schema registry's ID for a schema is globally unique, the binary
message encoding does not need to change, and consumers also handle the
mixture of record types without change. When a schema is changed, the
registry checks compatibility with previous schemas of the same
fully-qualified type name; different record types can be evolved
independently without any interference.