Introduce new properties to supply basic auth user/password
We introduce a new credential provider that is selected by using the
credential source "USERNAME_PASSWORD". The user must supply two options,
namely schema.registry.basic.auth.username and
schema.registry.basic.auth.password. The motivation for this additional
way is, that the user can provide the credentials in plain text, which
may be more convenient than percent-encoded credentials as part of the
target URL.