Fix conversion of anonymous option connect struct to avro
AvroConverter.fromConnectData first creates an avro schema for the passed object
and then creates an avro object from the struct and the generated schema. The
converter converts optional Schemas to avro unions where one entry is null and
the other is the converted non-optional schema. When converting non-optional
Structs to avro records, the converter assigns the record a name, choosing a
default if the Struct's schema has no name. Then, when filling in the data for
the optional field, the converter looks up the non-null schema in the union.
This look up tries to match the name to the name in the Connect schema, which
is null unless explicitly provided. So the lookup returns a union schema, which
you cannot create a GenericRecord from. The fix implemented in this patch is to
add special handling for null names when looking up the schema in the union.