CC-2769: Ability to optionally include in output the schema ID for key and value
A new `print.schema.ids=true` property can be used to print out the schema ID for the value’s schema, and the schema ID for the key iff the `print.key=true` is enabled. The `schema.id.separator` property can be used to change the delimiter used between the key/value and the schema ID; this defaults to the tab character (like the existing `key.separator`).

With `print.schema.ids=true`, the output might look similar to:
```
{"k":0} 1   {"f1":"value0"} 2
{"k":1} 1   {"f1":"value1"} 2
{"k":2} 1   {"f1":"value2"} 2
{"k":3} 1   {"f1":"value3"} 2
```
where `1` and `2` are the schema IDs for the keys and values, respectively.