CC-3803: Added support for logical schema defaults
Used the logical converters before attempting to morph the default literals used in Connect schemas into the primitive form needed for setting defaults on Avro schemas, and vice versa when morphing the Avro schema defaults to the logical literals needed by Connect schemas. Added to an existing test case to add logical type fields with defaults.

As part of this, changed the parameters for Decimal to only include the `connect.decimal.precision` schema property when it is not the default of 64. This makes it more consistent with repeated conversion and with what is originally generated from a Connect schema.