Reflection serde (#1281)
* enabled default constructor

Default constructor allows the SerDe to be used as default serde in KStreams config.

* added logic to resolve the reflection schema when it is not provided during construction

* fixed whitespace warning

* added missing constructors

* cloned the original test case to support dynamic type recognition

* cache only generated schemas

The failure in testVersionMaintained was caused by caching the schema for given name. In case
of GenericDatumReader the schema is as it comes from the serialized record. This means
if the schema evolves, there could be multiple representations of the same avro record.