fix: `connect.default` on long avro field
fixes: https://github.com/confluentinc/schema-registry/issues/1376

If an Avro schema has a `connect.default` property on a `long` field, e.g.

```json
{
    "name": "date_created",
    "type": [
      {
        "type": "long",
        "connect.version": 1,
        "connect.default": 0,
        "connect.name": "io.debezium.time.Timestamp"
      },
      "null"
    ],
    "default": 0
  }
```

Them `AvroData.toConnectSchema` throws:

```
org.apache.kafka.connect.errors.DataException: Invalid type for INT64: class java.lang.Integer
```

This bug was introduced in the upgrade to Avro 1.9.

The cause is that the call to get the value of `connect.default` was switched from `schema.getJsonProp`, which returned a IntNode, to `schema.getObjectProp`, which returns an int.  Later code that uses this default value differentiates between JsonNode's and non-JsonNodes.

* Non-JsonNodes are treated literally.  As the default is small enough to fit into an int, but the type is a long, the code throws the DataException.
* JsonNodes have their value extracted using type coercion, to ensure the default value matches the schema type.

The fix is to convert the returned default value back into a JsonNode so that the type coercion can happen.