DG-321 Omit schemaType field when storing Avro schemas
Schemas generated with SR 5.5 include a schemaType field, which prevents
SR 5.4 from reading them.  The fix is the same as with the REST APIs,
which is to store the Avro schema without a schemaType field (Protobuf
and JSON Schema will still have this field).  This will allow users to
downgrade from 5.5 to 5.4 and read any Avro schemas generated with 5.5
(of course Protobuf/JSON Schema schemas generated with 5.5 will still
not be able to be read by 5.4).