DG-399 Use envelope in JsonSchemaUtils.toObject
This method is mainly used by the REST Proxy and tests.
The return value is passed to KafkaJsonSchemaSerializer.
By wrapping the schema and payload in an envelope, the
serializer will use the schema from the envelope rather
than trying to generate one based on the JsonNode (which
will be incorrect).