Avro Console producer should support subject name strategy option
- Fixes #898 SchemaMessageReader (used by classes like
AvroMessageFormatter) no longer hard-codes the subject name format.
Instead, it leverages AbstractKafkaSchemaSerDeConfig's existing logic to
use the configuration properties to determine the subject name strategy
for the value and, if needed, the key. This allows use of the
"kafka-avro-console-producer" tool with  non-default subject name
strategies.
- Added unit test to exercise the logic from AvroMessageReader#init()
which was previously not covered.
- Spread MockSchemaRegistry functionality so it's not just in
AbstractKafkaSchemaSerDe, but can be used in other classes such as
SchemaMessageFormatter for more testability.
- Improved error message when schema not found to include the subject
name that was used.
- Fixed "Unsupported platform: protoc-3.11.4-osx-x86_64.exe" due to
using incorrect version of protoc-jar-maven-plugin.