Add kafkastore.topic.verify configurations options
Adds the `kafkastore.topic.verify` option (defaulting to `true` to
preserve current behaviour). This options controls whether
schema-registry performs verification and auto-creation of topics on
boot.

Disabling this functionality is required in environments where clients
do not have the `DescribeConfigs` or where topic auto creation is not
supported.

For example, this is required in order to run the schema-registry
with Heroku's managed Apache Kafka service (See [Kafka on Heroku ACLs]).

[Kafka on Heroku ACLs]: https://devcenter.heroku.com/articles/kafka-on-heroku#understanding-acls

Signed-off-by: Christian Gregg <christian@bissy.io>