Force non-required fields to be optional in JsonSchemaData
Fields can be "optional" in two ways:
- nullable
- not required

So a field in a Struct can be required and nullable, meaning the field
must exist in the struct but might not have a value. Non-required fields
in the schema were made nullable in the Connect schema but were still
required. This commit fixes that.